# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Public API for integrating with the crypto exchange platform.",
        "title": "Crypto Exchange API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/assets": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of all supported crypto and fiat assets along with their current live prices.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "List supported assets",
                "responses": {
                    "200": {
                        "description": "List of assets",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/wallet/swap": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Swap one asset for another internally",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Swap assets",
                "parameters": [
                    {
                        "description": "Swap Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwapRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwapResponse"
                        }
                    }
                }
            }
        },
        "/wallet/deposit/{assetID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Generates or retrieves a permanent deposit address for a specific asset ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Get deposit address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "assetID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deposit Address",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wallet/portfolio": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the developer's wallet portfolio including asset balances and their total USD value.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet portfolio",
                "responses": {
                    "200": {
                        "description": "Portfolio data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wallet/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Processes an outbound crypto transfer to an external address. Requires a secret key (sk_).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Send crypto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Current Unix timestamp in seconds",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature (timestamp + method + path + body)",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Send request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.SendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wallet/swap": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Executes an internal swap between two supported assets based on live prices. Requires a secret key (sk_).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Swap crypto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Current Unix timestamp in seconds",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HMAC-SHA256 signature (timestamp + method + path + body)",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Swap request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Swap details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wallet/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves paginated transaction history for the developer's account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "List transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transaction type (e.g. deposit, withdrawal, swap)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Swap": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "exchange_rate": {
                    "type": "string"
                },
                "fee_amount": {
                    "type": "string"
                },
                "from_amount": {
                    "type": "string"
                },
                "from_asset_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "to_amount": {
                    "type": "string"
                },
                "to_asset_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.SwapRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "The amount of from_asset to swap",
                    "type": "string"
                },
                "from_asset_id": {
                    "type": "integer"
                },
                "to_asset_id": {
                    "type": "integer"
                }
            }
        },
        "models.SwapResponse": {
            "type": "object",
            "properties": {
                "swap": {
                    "$ref": "#/definitions/models.Swap"
                }
            }
        },
        "services.SendRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "asset_id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "to_address": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}